feat(skeleton): Add new SkeletonTable docs
