+++ packages/module/patternfly-docs/content/extensions/component-groups/examples/SkeletonTable/SkeletonTableCustomExample.tsx
@@ -1,16 +1,15 @@
 import React from 'react';
-import SkeletonTable from '@patternfly/react-core/dist/js/components/Skeleton/SkeletonTable';
-import { Th } from '@patternfly/react-table';
+import SkeletonTable from "@patternfly/react-component-groups/dist/dynamic/SkeletonTable";import { Th } from '@patternfly/react-table';
 
 export const SkeletonTableExample: React.FC = () => (
   <SkeletonTable
-    rowSize={10}
+    rowsCount={10}
     columns={[
       <Th key="1" sort={{ columnIndex: 0, sortBy: {} }}>
-        first
+        First
       </Th>,
       <Th key="2" sort={{ columnIndex: 1, sortBy: {} }}>
-        second
+        Second
       </Th>
     ]}
   />
