+++ packages/module/patternfly-docs/content/extensions/component-groups/examples/SkeletonTable/SkeletonTableExpandableExample.tsx
@@ -1,0 +1,4 @@
+import React from 'react';
+import SkeletonTable from "@patternfly/react-component-groups/dist/dynamic/SkeletonTable";
+
+export const SkeletonTableExample: React.FC = () => <SkeletonTable rowsCount={10} columns={[ 'First', 'Second' ]} isExpandable />;
